+++ src/_gettsim/transfers/arbeitsl_geld_2/arbeitsl_geld_2_eink.py
@@ -3,7 +3,7 @@ from _gettsim.shared import policy_info
 
 
 def arbeitsl_geld_2_eink_m(
-    _arbeitsl_geld_2_nettoeink_ohne_transfers_m: float,
+    _arbeitsl_geld_2_nettoeink_mit_freibetrag_m: float,
     kind_unterh_erhalt_m: float,
     unterhaltsvors_m: float,
     kindergeld_zur_bedarfsdeckung_m: float,
@@ -21,8 +21,8 @@ def arbeitsl_geld_2_eink_m(
 
     Parameters
     ----------
-    _arbeitsl_geld_2_nettoeink_ohne_transfers_m
-        See :func:`_arbeitsl_geld_2_nettoeink_ohne_transfers_m`.
+    _arbeitsl_geld_2_nettoeink_mit_freibetrag_m
+        See :func:`_arbeitsl_geld_2_nettoeink_mit_freibetrag_m`.
     kind_unterh_erhalt_m
         See :func:`kind_unterh_erhalt_m`.
     unterhaltsvors_m
@@ -38,7 +38,7 @@ def arbeitsl_geld_2_eink_m(
 
     """
     return (
-        _arbeitsl_geld_2_nettoeink_ohne_transfers_m
+        _arbeitsl_geld_2_nettoeink_mit_freibetrag_m
         + kind_unterh_erhalt_m
         + unterhaltsvors_m
         + kindergeld_zur_bedarfsdeckung_m
@@ -46,16 +46,39 @@ def arbeitsl_geld_2_eink_m(
     )
 
 
-def _arbeitsl_geld_2_nettoeink_ohne_transfers_m(  # noqa: PLR0913
+def _arbeitsl_geld_2_nettoeink_mit_freibetrag_m(
+    arbeitsl_geld_2_nettoeink_m: float,
+    arbeitsl_geld_2_eink_anr_frei_m: float,
+) -> float:
+    """Net income after deductions for calculation of basic subsistence
+    (Arbeitslosengeld II / Bürgergeld).
+
+    Note: Since 2023, Arbeitslosengeld 2 is referred to as Bürgergeld.
+
+    Parameters
+    ----------
+    arbeitsl_geld_2_nettoeink_m
+        See :func:`arbeitsl_geld_2_nettoeink_m`.
+    arbeitsl_geld_2_eink_anr_frei_m
+        See :func:`arbeitsl_geld_2_eink_anr_frei_m`.
+
+    Returns
+    -------
+    Income after taxes, social insurance contributions, and other deductions.
+
+    """
+    return arbeitsl_geld_2_nettoeink_m - arbeitsl_geld_2_eink_anr_frei_m
+
+
+def arbeitsl_geld_2_nettoeink_m(
     arbeitsl_geld_2_bruttoeink_m: float,
     eink_st_y_sn: float,
     soli_st_y_sn: float,
     anz_personen_sn: int,
     sozialv_beitr_arbeitnehmer_m: float,
-    arbeitsl_geld_2_eink_anr_frei_m: float,
 ) -> float:
-    """Income (after deduction of taxes, social insurance contributions, and other
-    deductions) for calculation of basic subsistence.
+    """Net income for calculation of basic subsistence (Arbeitslosengeld II /
+    Bürgergeld).
 
     Note: Since 2023, Arbeitslosengeld 2 is referred to as Bürgergeld.
 
@@ -63,16 +86,14 @@ def _arbeitsl_geld_2_nettoeink_ohne_transfers_m(  # noqa: PLR0913
     ----------
     arbeitsl_geld_2_bruttoeink_m
         See :func:`arbeitsl_geld_2_bruttoeink_m`.
-    sozialv_beitr_arbeitnehmer_m
-        See :func:`sozialv_beitr_arbeitnehmer_m`.
     eink_st_y_sn
         See :func:`eink_st_y_sn`.
     soli_st_y_sn
         See :func:`soli_st_y_sn`.
     anz_personen_sn
         See :func:`anz_personen_sn`.
-    arbeitsl_geld_2_eink_anr_frei_m
-        See :func:`arbeitsl_geld_2_eink_anr_frei_m`.
+    sozialv_beitr_arbeitnehmer_m
+        See :func:`sozialv_beitr_arbeitnehmer_m`.
 
     Returns
     -------
@@ -84,7 +105,6 @@ def _arbeitsl_geld_2_nettoeink_ohne_transfers_m(  # noqa: PLR0913
         - (eink_st_y_sn / 12 / anz_personen_sn)
         - (soli_st_y_sn / 12 / anz_personen_sn)
         - sozialv_beitr_arbeitnehmer_m
-        - arbeitsl_geld_2_eink_anr_frei_m
     )